maw: get-all-events almost done
